Check if table exists before deleting data (important for deployment/migration).

+++ app/__init__.py
@@ -41,15 +41,16 @@ def create_app(config_name):
     # TODO: This  might not be the right place to do
     # initialization
     with app.app_context():
-        # Remove old data
-        models.Word.query.delete()
-        db.session.commit()
+        if db.engine.dialect.has_table(db.engine,'rjeci'):
+            # Remove old data
+            models.Word.query.delete()
+            db.session.commit()
     
-        # Populate the database from .csv
-        data = pd.read_csv('app/static/data/psiholeks.csv',
-                           delimiter=';',
-                           index_col='rijec')
-        data.to_sql('rijeci', db.engine, if_exists='append')
+            # Populate the database from .csv
+            data = pd.read_csv('app/static/data/psiholeks.csv',
+                               delimiter=';',
+                               index_col='rijec')
+            data.to_sql('rijeci', db.engine, if_exists='append')
         
     from .home import home as home_blueprint
     app.register_blueprint(home_blueprint)
